Initial Commit.
Added the tableview and a custom cell.
Added functionality to dynamically resize the cell once it is dismissed by the header, instead of it remaining there and only deleting the cells from the table.

ToDo:
Would like to refactor a little, and would like to place the reuse identifier inside the cell, or the table code.